+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+### Changes
+- ([\#74](https://github.com/forbole/juno/pull/74)) Applied `GetMissingHeights()` in `enqueueMissingBlocks()` & in `parse blocks missing` cmd
+
 ## v4.0.0
 ### Changes
 - Updated cosmos/cosmos-sdk to `v0.45.8`