+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Unreleased
-- ([\#88](https://github.com/forbole/juno/pull/88)) Added `DbLatestHeight` metric to prometheus monitoring
+
+### Changes
+- ([\#74](https://github.com/forbole/juno/pull/74)) Applied `GetMissingHeights()` in `enqueueMissingBlocks()` & in `parse blocks missing` cmd
+- ([\#88](https://github.com/forbole/juno/pull/88)) Added `juno_db_latest_height` metric to prometheus monitoring
 
 ## v4.0.0
 ### Changes